+++ Prototype/Assets/Scenes/TestMap.unity
@@ -119,52 +119,6 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!1 &416727075
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 416727077}
-  - component: {fileID: 416727076}
-  m_Layer: 0
-  m_Name: ShockwaveControl
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &416727076
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 416727075}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7fb1a261e63b91d42af2caabba821f42, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  pre_Shockwave: {fileID: 2041897516517486551, guid: fa0c6da7dcebd4143a7829715f7b712e, type: 3}
-  playerTransform: {fileID: 1860405353}
---- !u!4 &416727077
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 416727075}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.12059268, y: -0.23103279, z: 0.008982126}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &889125692
 GameObject:
   m_ObjectHideFlags: 0
@@ -447,9 +401,9 @@ GameObject:
   - component: {fileID: 1860405353}
   - component: {fileID: 1860405352}
   - component: {fileID: 1860405355}
-  - component: {fileID: 1860405358}
   - component: {fileID: 1860405357}
   - component: {fileID: 1860405356}
+  - component: {fileID: 1860405358}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -624,38 +578,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d6c0658c2b5739e43b4e42665b55b5d4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  ShockwaveSkill: {fileID: 416727076}
---- !u!1 &1914441426
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1914441427}
-  m_Layer: 0
-  m_Name: GameObject
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &1914441427
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1914441426}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.12059268, y: -0.23103279, z: 0.008982126}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  pre_Shockwave: {fileID: 2041897516517486551, guid: fa0c6da7dcebd4143a7829715f7b712e, type: 3}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
@@ -663,5 +586,3 @@ SceneRoots:
   - {fileID: 889125695}
   - {fileID: 1860405353}
   - {fileID: 1278911231}
-  - {fileID: 416727077}
-  - {fileID: 1914441427}
