--- Prototype/Assets/Scripts/Skills.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ae3af323b84984f418e32c8857bb0073
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
