--- Prototype/Assets/Scripts/Skills/ActivateShield.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class ActivateShield : SkillBase {
-    protected override void Execute() {
-
-
-
-    }
-}
