--- Prototype/Assets/Scripts/Skills/ActivateShield.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1092e2f336d85c3498ed463ace5320be
