--- Prototype/Assets/Scripts/Skills/PlayerSkillControl.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d6c0658c2b5739e43b4e42665b55b5d4
