--- Prototype/Assets/Scripts/Skills/Shockwave.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7fb1a261e63b91d42af2caabba821f42
