--- Prototype/Assets/Scripts/Skills/SkillBase.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6265b9631ab61a645b51588beb327750
